requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
googlesearch-python==1.2.5
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
pillow==11.1.0
py-redis==1.1.1
PyJWT==2.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
validators==0.34.0
Werkzeug==3.1.3
whois==1.20240129.2
